# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trading bot API for Trading 212
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/config:
    patch:
      operationId: updateBotConfig
      tags: [bot]
      summary: Update bot strategy configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConfigInput"
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
  /broker/connect:
    post:
      operationId: connectBroker
      tags: [broker]
      summary: Connect (or replace) this user's broker credentials
      description: >
        Tests the credentials against the broker before saving. Credentials are
        encrypted at rest and never returned to the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrokerConnectInput"
      responses:
        "201":
          description: Broker connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerStatus"
        "400":
          description: Invalid input or the credentials did not work
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantError"
...
  /instruments:
    get:
      operationId: listInstruments
      tags: [instruments]
      summary: List watchlist instruments
      responses:
        "200":
          description: Instrument list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instrument"
    post:
      operationId: addInstrument
      tags: [instruments]
      summary: Add instrument to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstrumentInput"
      responses:
        "201":
          description: Instrument added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrument"
  /instruments/{id}:
    delete:
      operationId: deleteInstrument
      tags: [instruments]
      summary: Remove instrument from watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /assistant/conversations:
    get:
      operationId: listConversations
      tags: [assistant]
      summary: List all assistant conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [assistant]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /assistant/conversations/{id}:
    get:
      operationId: getConversation
      tags: [assistant]
      summary: Get a conversation with its messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantError"
    delete:
      operationId: deleteConversation
      tags: [assistant]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantError"
  /assistant/conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [assistant]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssistantMessage"
    post:
      operationId: sendMessage
      tags: [assistant]
      summary: Send a message and receive a streaming assistant response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
...
  /charts/insight:
    get:
      operationId: getChartInsight
      tags: [charts]
      summary: AI technical read for an instrument (trend, S/R, volatility, confidence)
      parameters:
        - name: epic
          in: query
          required: true
          schema:
            type: string
        - name: resolution
          in: query
          schema:
            type: string
            default: HOUR
      responses:
        "200":
          description: Chart insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartInsight"
        "400":
          description: Missing epic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantError"
        "502":
          description: Failed to compute insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantError"
...
components:
  schemas:
...
    BotStatus:
      type: object
      required: [running, config, circuitBreaker]
      properties:
        running:
          type: boolean
        lastRunAt:
          type: ["string", "null"]
        nextRunAt:
          type: ["string", "null"]
        config:
          $ref: "#/components/schemas/BotConfig"
        circuitBreaker:
          $ref: "#/components/schemas/CircuitBreaker"
    CircuitBreaker:
      type: object
      required: [tripped]
      description: Daily-loss circuit breaker state. When tripped, the bot is stopped and will not trade until manually resumed.
      properties:
        tripped:
          type: boolean
          description: True when the daily loss limit has been hit and trading is halted.
        reason:
          type: ["string", "null"]
          description: Human-readable explanation of why the breaker tripped.
        trippedAt:
          type: ["string", "null"]
          description: ISO timestamp when the breaker tripped.
        dayStartEquity:
          type: ["number", "null"]
          description: Account total equity recorded at the start of the current UTC day, used as the loss baseline.
    BotConfig:
      type: object
      required: [shortPeriod, longPeriod, tradeAmount, intervalMinutes, dryRun, broker, stopLossPercent, takeProfitPercent, riskPerTradePercent, maxPositionSizePercent, maxDailyLossPercent, maxConcurrentPositions, aiTradeMode, regimeFilterEnabled, costPerTradePercent]
      properties:
        shortPeriod:
          type: integer
          description: Short MA period (e.g. 9)
        longPeriod:
          type: integer
          description: Long MA period (e.g. 21)
        tradeAmount:
          type: number
          description: Fixed amount in account currency per trade (ignored when riskPerTradePercent > 0)
        intervalMinutes:
          type: integer
          description: How often to run signal check (minutes)
        dryRun:
          type: boolean
          description: If true, log signals but do not place orders
        broker:
          type: string
          enum: [trading212, capitalcom]
          description: Which broker to route trades through
        stopLossPercent:
          type: number
          description: Stop-loss distance as % of entry price (e.g. 2 = 2%). 0 disables stop loss.
        takeProfitPercent:
          type: number
          description: Take-profit distance as % of entry price (e.g. 4 = 4%). 0 disables take profit. Capital.com only — ignored on Trading 212.
        riskPerTradePercent:
          type: number
          description: Account balance % to risk per trade for position sizing (e.g. 1 = 1%). 0 uses fixed tradeAmount.
        maxPositionSizePercent:
          type: number
          description: Hard cap on a single position's value as % of account balance (e.g. 5 = 5%). Position size is clamped to this. 0 disables the cap.
        maxDailyLossPercent:
          type: number
          description: Daily-loss circuit breaker threshold as % of the day-start equity (e.g. 3 = 3%). When reached, the bot stops and must be manually resumed. 0 disables the breaker.
        maxConcurrentPositions:
          type: integer
          description: Maximum number of simultaneously open positions. New entries (long or short) are blocked at this limit; trades on an already-open ticker are unaffected. 0 disables the cap.
        aiTradeMode:
          type: string
          enum: [off, guard, autonomous]
          description: How Claude participates in execution. off = strategy only; guard = Claude approves/vetoes each MA signal; autonomous = Claude decides trades.
        regimeFilterEnabled:
          type: boolean
          description: When true, each instrument is classified trending/ranging (close-based ADX) and routed to trend-following or mean-reversion automatically. When false, only trend-following runs.
        costPerTradePercent:
          type: number
          description: Estimated round-trip trading cost (spread + commission) as a % of trade value, e.g. 0.1 = 0.1%. Used by the backtester for cost-aware expectancy. 0 assumes frictionless trades. Does not affect live orders.
    BotConfigInput:
      type: object
      properties:
        shortPeriod:
          type: integer
        longPeriod:
          type: integer
        tradeAmount:
          type: number
        intervalMinutes:
          type: integer
        dryRun:
          type: boolean
        broker:
          type: string
          enum: [trading212, capitalcom]
        stopLossPercent:
          type: number
          description: Stop-loss distance as % of entry price. 0 disables.
        takeProfitPercent:
          type: number
          description: Take-profit distance as % of entry price. 0 disables. Capital.com only.
        riskPerTradePercent:
          type: number
          description: Account balance % to risk per trade. 0 uses fixed tradeAmount.
        maxPositionSizePercent:
          type: number
          description: Hard cap on a single position's value as % of account balance. 0 disables.
        maxDailyLossPercent:
          type: number
          description: Daily-loss circuit breaker threshold as % of day-start equity. 0 disables.
        maxConcurrentPositions:
          type: integer
          description: Maximum number of simultaneously open positions. New entries (long or short) are blocked at this limit; trades on an already-open ticker are unaffected. 0 disables the cap.
        aiTradeMode:
          type: string
          enum: [off, guard, autonomous]
          description: How Claude participates in execution.
        regimeFilterEnabled:
          type: boolean
          description: Enable automatic trending/ranging routing between trend-following and mean-reversion.
        costPerTradePercent:
          type: number
          description: Estimated round-trip trading cost (spread + commission) as a % of trade value. 0 assumes frictionless trades. Backtest-only; does not affect live orders.
    BrokerConnectInput:
      type: object
      required: [broker]
      properties:
        broker:
          type: string
          enum: [trading212, capitalcom]
        capital:
          type: object
          description: Required when broker is capitalcom.
          properties:
            apiKey:
              type: string
            identifier:
              type: string
            password:
              type: string
        trading212:
          type: object
          description: Required when broker is trading212.
          properties:
            apiKey:
              type: string
    BrokerStatus:
      type: object
      required: [connected]
      properties:
        connected:
          type: boolean
        broker:
          type: string
          enum: [trading212, capitalcom]
        identifierMasked:
          type: string
          description: Partially masked identifier/key for display — never the credential itself.
...
    Instrument:
      type: object
      required: [id, ticker, name, enabled]
      properties:
        id:
          type: integer
        ticker:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        addedAt:
          type: string
    InstrumentInput:
      type: object
      required: [ticker, name]
      properties:
        ticker:
          type: string
        name:
          type: string
        enabled:
          type: boolean
...
    Conversation:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
    AssistantMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    ConversationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    ConversationWithMessages:
      type: object
      required: [id, title, createdAt, messages]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AssistantMessage"
    AssistantError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ChartInsight:
      type: object
      required:
        - epic
        - trend
        - volatility
        - confidence
        - explanation
        - riskWarning
      properties:
        epic:
          type: string
        trend:
          type: string
          enum: [Uptrend, Downtrend, Sideways]
        support:
          type: ["number", "null"]
        resistance:
          type: ["number", "null"]
        volatility:
          type: string
          enum: [Low, Medium, High]
        confidence:
          type: integer
        explanation:
          type: string
        riskWarning:
          type: string
...